feat: 뉴스스탠드 헤더의 상수 파일 분리
