+++ src/components/news-stand-header.js
@@ -1,17 +1,4 @@
-const ICON_NEWS = './assets/icons/newspaper.svg';
-const HEADER_CLASS = {
-  WRAPPER: 'header__wrapper',
-  BUTTON: 'header__button',
-  IMAGE: 'header__img',
-  H1: 'header__h1',
-  TIME: 'header__time',
-};
-const DATE_OPTIONS = {
-  year: 'numeric',
-  month: '2-digit',
-  day: '2-digit',
-  weekday: 'long',
-};
+import { HEADER_CLASS, DATE_OPTIONS, PATH } from '../constants/news-stand-header.js';
 
 /**
  * 뉴스스탠드의 INIT
@@ -27,7 +14,7 @@ const setHeader = () => {
   const headerWrapper = document.querySelector(`.${HEADER_CLASS.WRAPPER}`);
   const headerElement = `
     <button class=${HEADER_CLASS.BUTTON}>
-      <img class=${HEADER_CLASS.IMAGE} src=${ICON_NEWS}></img>
+      <img class=${HEADER_CLASS.IMAGE} src=${PATH.LOGO}></img>
       <h1 class=${HEADER_CLASS.H1}>뉴스스탠드</h1>
     </button>
     <time class=${HEADER_CLASS.TIME}></time>
