+++ src/constants/news-stand-header.js
@@ -1,0 +1,19 @@
+const HEADER_CLASS = {
+  WRAPPER: 'header__wrapper',
+  BUTTON: 'header__button',
+  IMAGE: 'header__img',
+  H1: 'header__h1',
+  TIME: 'header__time',
+};
+const DATE_OPTIONS = {
+  year: 'numeric',
+  month: '2-digit',
+  day: '2-digit',
+  weekday: 'long',
+};
+
+const PATH = {
+  LOGO: './assets/icons/newspaper.svg',
+};
+
+export { HEADER_CLASS, DATE_OPTIONS, PATH };
